+++ specs/configuration.spec.php
@@ -56,7 +56,7 @@ describe('Configuration', function() {
             $this->configuration->setSkipPattern('/skip/');
             $this->configuration->setGrep('*.test.php');
             $this->configuration->setReporter('reporter');
-            $this->configuration->setPath('/tests');
+            $this->configuration->setPaths(['/tests-a', '/tests-b']);
             $this->configuration->disableColors();
             $this->configuration->stopOnFailure();
             $this->configuration->setDsl(__FILE__);
@@ -67,6 +67,7 @@ describe('Configuration', function() {
             $grep = getenv('PERIDOT_GREP');
             $reporter = getenv('PERIDOT_REPORTER');
             $path = getenv('PERIDOT_PATH');
+            $paths = getenv('PERIDOT_PATHS');
             $colors = getenv('PERIDOT_COLORS_ENABLED');
             $stop = getenv('PERIDOT_STOP_ON_FAILURE');
             $dsl = getenv('PERIDOT_DSL');
@@ -76,7 +77,8 @@ describe('Configuration', function() {
             assert($skipPattern === '/skip/', 'should have set skip pattern env');
             assert($grep === '*.test.php', 'should have set grep env');
             assert($reporter === 'reporter', 'should have set reporter env');
-            assert($path === '/tests', 'should have set path env');
+            assert($path === '/tests-a', 'should have set path env');
+            assert($paths === '/tests-a' . PATH_SEPARATOR . '/tests-b', 'should have set paths env');
             assert(!$colors, 'should have set colors env');
             assert($stop, 'should have set stop env');
             assert($dsl === __FILE__, 'should have set dsl env');
@@ -121,6 +123,34 @@ describe('Configuration', function() {
 
             assert(preg_match($pattern, 'lmao lol(wat huh'), 'normalized pattern should match text with prefixes and suffixes');
             assert(!preg_match($pattern, 'lolol(wat'), 'normalized pattern should not match text without word boundaries');
+        });
+    });
+
+    describe('->setPath()', function() {
+        it('should set both path and paths', function() {
+            $this->configuration->setPath('/tests');
+
+            assert($this->configuration->getPath() === '/tests', 'should have set path');
+            assert($this->configuration->getPaths() === ['/tests'], 'should have set paths');
+        });
+    });
+
+    describe('->setPaths()', function() {
+        it('should set both path and paths', function() {
+            $this->configuration->setPaths(['/tests-a', '/tests-b']);
+
+            assert($this->configuration->getPath() === '/tests-a', 'should have set path');
+            assert($this->configuration->getPaths() === ['/tests-a', '/tests-b'], 'should have set paths');
+        });
+
+        it('should disallow setting an empty paths array', function() {
+            $exception = null;
+            try {
+                $this->configuration->setPaths([]);
+            } catch (InvalidArgumentException $e) {
+                $exception = $e;
+            }
+            assert(!is_null($exception), 'expected exception to be thrown');
         });
     });