+++ specs/input-definition.spec.php
@@ -8,13 +8,6 @@ describe('InputDefinition', function() {
         $this->definition = new InputDefinition();
     });
 
-    describe('->argument()', function() {
-        it('should add an argument to the definition', function() {
-            $this->definition->argument('myarg', InputArgument::OPTIONAL, 'an arg');
-            assert(!is_null($this->definition->getArgument('myarg')), 'argument() should register argument');
-        });
-    });
-
     describe('->option()', function() {
         it('should add an option to the definition', function() {
             $this->definition->option('myopt','-m', InputOption::VALUE_NONE, 'an opt');