+++ src/Configuration.php
@@ -40,9 +40,9 @@ class Configuration
     protected $reporter = 'spec';
 
     /**
-     * @var string
+     * @var array
      */
-    protected $path;
+    protected $paths;
 
     /**
      * @var string
@@ -61,7 +61,7 @@ class Configuration
 
     public function __construct()
     {
-        $this->path = getcwd();
+        $this->paths = [getcwd()];
         $this->configurationFile = getcwd() . DIRECTORY_SEPARATOR . 'peridot.php';
         $this->dsl = __DIR__ . DIRECTORY_SEPARATOR . 'Dsl.php';
     }
@@ -158,7 +158,7 @@ class Configuration
      */
     public function setPath($path)
     {
-        return $this->write('path', $path);
+        return $this->writePaths([$path]);
     }
 
     /**
@@ -168,7 +168,32 @@ class Configuration
      */
     public function getPath()
     {
-        return $this->path;
+        return $this->paths[0];
+    }
+
+    /**
+     * Set the paths to load tests from
+     *
+     * @param array $paths
+     * @return $this
+     */
+    public function setPaths(array $paths)
+    {
+        if (empty($paths)) {
+            throw new \InvalidArgumentException('Paths cannot be empty.');
+        }
+
+        return $this->writePaths($paths);
+    }
+
+    /**
+     * Return the paths being searched for tests
+     *
+     * @return array
+     */
+    public function getPaths()
+    {
+        return $this->paths;
     }
 
     /**
@@ -295,8 +320,8 @@ class Configuration
      * Write a configuration value and persist it to the current
      * environment.
      *
-     * @param $varName
-     * @param $value
+     * @param string $varName
+     * @param string $value
      * @return $this
      */
     protected function write($varName, $value)
@@ -305,6 +330,20 @@ class Configuration
         $parts = preg_split('/(?=[A-Z])/', $varName);
         $env = 'PERIDOT_' . strtoupper(join('_', $parts));
         putenv($env . '=' . $value);
+        return $this;
+    }
+
+    /**
+     * Write the paths and persist them to the current environment.
+     *
+     * @param array $paths
+     * @return $this
+     */
+    protected function writePaths(array $paths)
+    {
+        $this->paths = $paths;
+        putenv('PERIDOT_PATH=' . $paths[0]);
+        putenv('PERIDOT_PATHS=' . implode(PATH_SEPARATOR, $paths));
         return $this;
     }