+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next release
 
+- **[BC BREAK]** Support for specifying multiple paths via command line
+  arguments ([#200])
 - **[NEW]** Introduced focused specs for powerful test isolation ([#199],
   [#204], [#197], [#194], [#188], [#185], [#181])
 - **[IMPROVED]** Stack trace output now excludes irrelevant information ([#203],
@@ -19,6 +21,7 @@
 [#197]: https://github.com/peridot-php/peridot/pull/197
 [#198]: https://github.com/peridot-php/peridot/pull/198
 [#199]: https://github.com/peridot-php/peridot/pull/199
+[#200]: https://github.com/peridot-php/peridot/pull/200
 [#201]: https://github.com/peridot-php/peridot/pull/201
 [#203]: https://github.com/peridot-php/peridot/pull/203
 [#204]: https://github.com/peridot-php/peridot/pull/204